students list  get with fillters

+++ food-order-managment-client/src/app/students/page.tsx
@@ -1,9 +1,9 @@
 "use client"
-import React, { useRef, useState } from 'react'
+import React, { useEffect, useRef, useState } from 'react'
 import Header from '../components/Header'
 import TableHeader from '../table/TableHeader'
 import tableInterface from '../table/TableInteface'
-import { Edit, Trash2 } from 'lucide-react';
+import { Edit, Search, Trash2 } from 'lucide-react';
 import TableRow from '../table/TableRow'
 import { useMutation, useQuery } from '@tanstack/react-query'
 import axios, { Axios } from 'axios'
@@ -33,8 +33,6 @@ const tableHeader: tableInterface[] = [
         style: ""
     }
 ]
-
-
 
 
 
@@ -92,129 +90,86 @@ const tableHeader: tableInterface[] = [
 
 
 
+
 const Students = () => {
 
     const [currentPage, setCurrentPage] = useState(1);
     const [totalPages, setTotalPages] = useState(1);
+    const [isAllApprovedusers, setIsAllApprovedUsers] = useState(false);
     const [data, setData] = useState([]);
 
+    let tableRows: any = []
 
-    // const TableRow=data.map((obj:any)=>{
-    //     style:"",
-    //     cellData:[
-    //         { isButton: "", text: obj.id, style: "" },
-    //         { isButton: "", text: obj.name, style: "" },
-    //         { isButton: "", text: `$${obj.price}`, style: "py-6" },
-    //         { isButton: obj.status, text: obj.status, style: "px-3 py-2" },
-    //     ]
-    // })
-
-    let tableRows = []
-    //     {
-    //         style: "",
-    //          cellData:[{
-    //             isButton: 'Complete',
-    //             text: 'jdojfo',
-    //             style: ''
-    //         },
-    //         {
-    //             isButton: '',
-    //             text: '1299',
-    //             style: 'py-8'
-    //         }
-    //             ,
-    //         {
-    //             isButton: '',
-    //             text: '2323',
-    //             style: ''
-    //         }
-    //             , {
-    //             isButton: '',
-    //             text: '121',
-    //             style: ''
-    //         },
-    //         {
-    //             isButton: 'icons',
-    //             text: <Trash2 />,
-    //             click: deleteItemFromArray,
-    //             style: 'px-3'
-    //         }]
-    //     },
-        // {
-        //     id: 3,
-        //     phoneNumber: 987654321,
-        //     studentName: "Jane Smith",
-        //     status: "Inactive",
-        //     actions:<Trash2/>
-        // },
-        // {
-        //     id: 3,
-        //     phoneNumber: 987654321,
-        //     studentName: "Jane Smith",
-        //     status: "Inactive",
-        //     actions:<Trash2/>
-        // }
-    // ]
 
     const [students, setStudents] = useState([]);
-    let RefObj=useRef(-1);
-    tableRows=students.map((obj:any,index:number)=>{
-        
-        return{style:'',
-            cellData:[
-              {  isButton:'',
-                text:obj.name,
-                style:''
-              },
-              {
-                isButton:'',
-                text:obj.username,
-                style:''
-              },
-              {
-                isButton:'',
-                text:obj.roles[0].name,
-                style:''
-              },
-              {
-                isButton:[obj.roles[0].name=="ROLE_STUDENT"?'Complete':obj.roles[0].name=="ROLE_STAFF"?'Pending':'Rejected'],
-                text:obj.status,
-                style:''
-              },
-              {
-                isButton:'icon',
-                text:[<Trash2 onClick={()=>deleteFuction(obj.id)}/>,<Edit/>],
-                style:'flex flex-row gap-4'
-              }
+    let [ref, setRef] = useState(-1);
+
+    tableRows = students.map((obj: any, index: number) => {
+
+        return {
+            style: '',
+            cellData: [
+                {
+                    isButton: '',
+                    text: obj.name,
+                    style: ''
+                },
+                {
+                    isButton: '',
+                    text: obj.username,
+                    style: ''
+                },
+                {
+                    isButton: '',
+                    text: obj.roles[0].name,
+                    style: ''
+                },
+                {
+                    isButton: [obj.roles[0].name == "ROLE_STUDENT" ? 'Complete' : obj.roles[0].name == "ROLE_STAFF" ? 'Pending' : 'Rejected'],
+                    text: obj.status,
+                    style: ''
+                },
+                {
+                    isButton: 'icon',
+                    text: [<Trash2 onClick={() => deleteFuction(obj.id)} />, <Edit />],
+                    style: 'flex flex-row gap-4'
+                }
 
             ]
-        
+
         }
     })
-    const deleteMutation=useMutation({
-        mutationFn:deleteFromDatabase(RefObj.current),
-        mutationKey:[students],
-        
+
+
+    const deleteMutation = useMutation({
+        mutationFn: deleteFromDatabase(ref),
+        mutationKey: [students],
+
     })
-    
-    function deleteFuction(index: number){
-        RefObj.current=index;
-        deleteMutation.mutate()
-        students.pop();
+
+    function deleteFuction(index: number) {
+        setRef(index);
+        if (ref != -1) {
+            deleteMutation.mutate()
+            students.pop();
+        }
     }
     useQuery({
-        queryKey: [students, 'allStudents'],
+        queryKey: ['allStudents'],
         queryFn: () => fetchDataForAllStudents(currentPage - 1)
     })
     function fetchDataForAllStudents(pageNum: any) {
+
         axios.get(`${process.env.NEXT_PUBLIC_BASE_URL}/api/users/getAll?page=${pageNum}&size=10`)
             .then((res: any) => {
                 setStudents(res?.data?.content)
+                setTotalPages(res?.data?.totalPages)
                 return res
             })
-            .catch((ex:any)=>{
+            .catch((ex: any) => {
                 return ex;
             })
+
     }
     console.log(students)
     const handlePrevious = () => {
@@ -228,13 +183,59 @@ const Students = () => {
         }
     }
 
-    function getStudentsForApproval(event:any): void {
-        
+    const studentMutaion = useMutation(
+        {
+            mutationKey: [students],
+            mutationFn: () => getStudentWithStatus(currentPage - 1)
+        }
+    )
+
+    function getStudentsforAppoval(): any {
+        setIsAllApprovedUsers((prev) => !prev)
+        studentMutaion.mutate();
     }
 
-    // function deleteFromDatabase(index: any){
-    //    axios.
-    // }
+    function getStudentWithStatus(page: number): any {
+        if (isAllApprovedusers) {
+            let object = {
+                type: "piriven",
+                status: "PENDING"
+            }
+            axios.post(`${process.env.NEXT_PUBLIC_BASE_URL}/api/users/type?page=${page}`, object).then((res: any) => {
+                setStudents(res?.data?.content)
+                setTotalPages(res?.data?.totalPages)
+                return res
+            });
+        } else {
+            fetchDataForAllStudents(0);
+        }
+    }
+
+
+   const serachStudentMutaion = useMutation({
+        mutationKey: ['searchStudent'],
+        mutationFn: () => fetchDataForSearchedStudents(currentPage - 1,sarch)
+    })
+
+    function fetchDataForSearchedStudents(page: number,userName:any): any {
+        axios.get(`${process.env.NEXT_PUBLIC_BASE_URL}/api/users/serach?username=${userName}&pageable=${page}`)
+        .then((res)=>{
+        console.log(res);
+            setStudents(res?.data?.content)
+            return res
+        }).catch((err)=>{
+            setStudents([])
+            return err
+        })
+    }
+
+    function getStudentsList(): any {
+        serachStudentMutaion.mutate();        
+    }
+    const [sarch,setSearch]:any = useState('');
+    function search(e:any):any{
+        setSearch(e.target.value);
+    }
 
     return (
         <div className="w-screen h-screen flex flex-col max-lg:w-2/3">
@@ -244,19 +245,29 @@ const Students = () => {
             <div className="w-full flex flex-col items-center">
 
                 <div className="p-6 w-[85%] bg-white   rounded-xl shadow-md  ">
-                    <h3 className="font-sans text-3xl font-semibold px-3 py-4 mb-24 ">View Students</h3>
+                    <h3 className="font-sans text-3xl font-semibold px-3 py-4 mb-10 ">View Students</h3>
 
                     <div className='flex items-end gap-5'>
                         <button className=' px-8 py-4 rounded-2xl bg-[#e6f6e9] font-sans font-semibold mb-10  '>Register Students</button>
-                        <button className=' px-8 py-4 rounded-2xl bg-[#e6e6f6] font-sans font-semibold mb-10  ' onClick={getStudentsForApproval}>Approve Students</button>
+                        <button className=' px-8 py-4 rounded-2xl bg-[#e6e6f6] font-sans font-semibold mb-10  ' onClick={() => getStudentsforAppoval()}>{isAllApprovedusers ? 'Approve Students' : 'All Students'}</button>
+                        <div className=' m-12  ring  rounded border'>
+                            <input
+                                type="text"
+                                name="search"
+                                placeholder="Search..."
+                                onChange={(e)=> search(e)}
+                            />
+                        <button className=' px-8 py-4 rounded-2xl bg-[#e6e6f6] font-sans font-semibold mb-10  ' onClick={() => getStudentsList()}> Search Student</button>
+
+                        </div>
                     </div>
 
                     <table className='w-full border-collapse'>
                         <TableHeader header={tableHeader} />
 
-                        {tableRows.map((row:any,index:number)=>{
-                            return(
-                               <TableRow key={index} cellData={row.cellData} styles={row.style}></TableRow>
+                        {tableRows.map((row: any, index: number) => {
+                            return (
+                                <TableRow key={index} cellData={row.cellData} styles={row.style}></TableRow>
                             )
                         })}
 
@@ -290,10 +301,11 @@ const Students = () => {
 export default Students
 
 
-function deleteFromDatabase(indexforDelete: number){
-    console.log(indexforDelete)
-  axios.delete(`${process.env.NEXT_PUBLIC_BASE_URL}/api/users/delete/${indexforDelete}`).then((res:any)=>{
-    console.log(res)
-  })
+function deleteFromDatabase(indexforDelete: number): any {
+    if (indexforDelete !== -1) {
+        axios.delete(`${process.env.NEXT_PUBLIC_BASE_URL}/api/users/delete/${indexforDelete}`).then((res: any) => {
+            console.log(res)
+        })
+    }
 }
